+++ src/main/java/com/woowacourse/support/exception/ErrorResponse.java
@@ -17,4 +17,8 @@ public class ErrorResponse {
     public static ErrorResponse from(final RuntimeException exception) {
         return new ErrorResponse(exception.getMessage());
     }
+
+    public static ErrorResponse from(final String message) {
+        return new ErrorResponse(message);
+    }
 }
